[FIX] available quantity computing
In order to filter on kind of bin a domain was created by filtering on
location ids. This request was badly translated by odoo checking for
parent_left and parent_right for each location to find children.
This because the research is not meant to search by leaves.

In order to simplify the request we add the capability to filter
available quantity directly by location kind. Thus we save a lot of
memory.

An other option would be to ensure all bin locations are in the same
tree and filter by the parent location for all bin.