Create user for WSO2
Create a user which has been communicated to Smile.  Smile will use this
user from WSO2 to send requests to Odoo controllers.  We only need a
public user, as they will only use a few custom controllers, so the user
doesn't count in the Odoo licensing. Those controllers will then use
sudo() to be able to work with more privileges.

Password is in lastpass.