Generalize the 'mark as exported'
Product and customers both have a 'esb_exported' flag. This flag must be
set to True when they are exported to the ESB. Only the one on the
product was set to True. Now this is a generic feature which can be
activated by setting '_mark_as_exported = True' in the exporter.

Fixes along the way:

* When no record was exported, the UPDATE query was crashing
* The field product_template.esb_exported is a computed field which is
  True if at least one variant has the 'esb_exported' mark. We write on
  the product's esb_exported field using a direct SQL query so we don't
  change the 'write_date' which would trigger a new export. Which means
  we the computed field is never triggered. Now this field is manually
  updated with SQL as well.