Remove import of lots from DB2
The lots are not required on migrated sale and purchase order
those were imported to fullfill the default requirement of product with lots.
Lots will be assigned later with inventory.

This should fix most errors with lot about quantities and duplicates
during import.

Checks on tracking are disabled trought context.