Delay only one export job on sales orders changes
Several jobs are create because of the way writes are called and events
triggered:

* a sale.order is created -> create triggers an export job
* a write is called on this sale.order by mail.thread -> triggers a
second export job
* a write is called on each line -> a job is created for each line

But all of these jobs are exporting the same sale.order. If we went
through the write or create of a sale, we know that lines should not
trigger jobs. If we went through a create, we know that we should
ignore a subsequent write on the same sale.

An alternative correction was proposed here:
https://github.com/camptocamp/alcyon_odoo/pull/637

The main differences are:

* The other proposal has to search in the database if a similar job
doesn't exist each time a new job is delayed, while here it's only based
on the state of the context.
* This proposal fixes mainly the case where a create/write is done on
the 'sale.order' model, which is always the case from the UI. If lines
are often changed in batch somewhere in the code, directly on
sale.order.line and not through sale.order, we'll still have duplicate
jobs. In that case we may want to use the other proposal as well (they
can be complementary).

ALCN-1392