ALCN-1415 : Fix possible missing records in esb export
If a record is being updated in a transaction that started before the
export but is not finished. The newly updated record would have a
write_date anterior to the last_export date on the timestamp.
And would miss the next export.
Also add queue_job_cron and set all cron jobs from the ESB to be run
through a queue job.