Change lock mechanism when exporting records (ESB)
The lock system on the records when they are being exported with
FOR UPDATE NO WAIT did not succeed at all on the product.product table when
addressing many records. So it is disabled and to be sure that we don't
miss a record that was updated during the last export, 5 minutes are
subtracted from the timestamp on each export. Poor man locking system !