Fix picking states created by imported orders
This solves 2 issues. The first is the wizard on purchase
for 'stock.backorder.choice' that broke the do_transfer for receptions.
In this case we bypass the wizard to do a standard transfer.

The second issue was picking in assigned state for both sales and
purchases. Those pickings creates quants and messup the stock
with negative quantities.
Here we unreserve the moves which removes the quants.