orders import - delete lines on update
Base data has now a 'deleted' column.
This base data table in postgres is a replication of AS400 table.

The 'deleted' column values are filled by an external script.

For new orders, deleted lines will be skipped.

For old orders, this adds first a verification for all order lines
if it was deleted based on this 'deleted' column.
If so, we force the order to draft. Which will trigger a recomputation
of the order when orders in draft are recomputed.

This change also include the possibility to recompute draft order in
'history' mode. Which may help to close in advance few Orders that
are over 120 days old.

Now, recomputation of drafts will always be called on 'final' mode.