Fix import of closed POs that were creating unwanted pickings
When a PO was already in the database with purchase state it was creating a picking
even before the importer is choosing to create one or not.
This is due to core `purchase` module. It creates a picking when a purchase order line
is written and field product_qty is set.

To avoid it, we set the state of the purchase to draft before making any change on the lines.