Fix wich records are exported in the customer flux
Following the change with 'mushroom' customer, some res_partner with
a parent_id must be exported as well.
Exporting all record of type customer, because there is not a simple
way to differentiate between the addresses of a simple customer and
the subsidiary of a head quarter.
For mushroom customer they have a parent_id and if it is a record of
type 'invoice' they are delivered and invoiced at this address. If
they have a type 'delivery', they are delivered at that address but
the invoicing is done to the address of the head quarter.