Add warning on sale exeception
ALCYN-1487

Allow for some sale exception to be warning only. An exception warning
will not be blocking for the confirmation of a sale order but when
raised it will add a pre-configured message (if defined) to the description of the
sale order line.
Also reactivate the detection of sale exception on the sale order record
and improve the styling of the displayed exception.