Create a fake migration location for SO import
On import of sale orders when we had partial delivery
we were creating picking and moves that when creating
negative quants on different locations affecting the
state of the stock.

To avoid this, we create a fake location only for imported Sales orders.
And when a sale order is imported and partially delivered, all done
quantities on AS400 side are marked as comming from that fake location.

ALCYN-1500