SO import - Group shipping backorders by partners
Only the backorder shipping must be grouped to avoid having to make
extra shippings for same customer after data import.

However, already sent shipping quantities remains in separate
done shippings for the process of validating them to go correctly.

Thus we still disable the auto-regroup of module `stock_groupbypartner`
and we merge the pickings at the end of import.