Create migration invoices on sales for partial deliveries
This adds invoices on imported sales.

It creates a migration journal for sales and a second one for purchases
Those invoices are only there to compute the `qty_invoiced` field.

This is why we set them on a special account.journal + we set the prices
to zero.

ALCYN-1525

This also change the migration purchase invoices to have the same
properties, zero prices and special journal.

ALCYN 1508