Fix productType in customer form statistique (ESB)
ALCYN-1543

The param recieved to filter on the product type does not use the
esb_ref value but full words (e.g., materiel is used for MAT).

And for the return value the productType must be the one of the business
unit, basically should be ('MAT', 'MED', 'ALI') translated to full words
like explained above.