ALYN-1622: Load New Food warehouse setup
ALCYN-1622: In the warehouse, products in the first rows are not associated to a dedicated bin anymore but are associated to a row. This is managed at reception by putting the products on the right rows.
Each row has its own parking (and unique) in front of the row.
For this, I removed the stock putaway route (stock > bin) and added a parking putaway route (parking > parking row).
So at reception, instead of freely choosing a parking, they are now directed to a dedicated parking.

Also include correction of some wrong setup:

    ALCYN-1632: Products are not sent to the parking according to the zone forced on the product

Split locations by rows for Food

Remove locator from some "aliments" and add a reception rule to push to right parking. Adapt migration to support this new setup.

Add L and Z avenue

Fix putaway xmlids and replay setup

migration: import changes

Update CSV file