Fix product import on PIM warning_info too long
ALCYN-1631

Products can not be imported on the PIM when the warning_info field is
longer than 255 characters.
So trimming the filed content. It is not such a nice solution but with
no feedback on the problem, should at least allow for the products to be
present on the PIM.