Fix quantity at zero on so line with exception (ESB).
ALCYN-1709

The quantity of a sale order line is set to zero if the line has a sale
exception on it.
It has been discussed that a sale order coming from Magento should be
trusted, so sale exception can be ignored.
But for sale orde coming from New Pharma (direct call to the api) if
a line as a sale exception it is set to zero.