Fix Via Pharmacy on export product (ESB)
ALCYN-1725

The field gescgr exported to the ESB can not be calculated for all
the products anymore because it has not been imported into Odoo
from db2, following the restructuration of product categories into a
tree structure.
So it is deducted for some specific product categories when needed by
the ESB. Otherwise it is left at zero.