Fix test on new separation of trusted vs untrusted
Fix test and add a test case.

With the use of webservice,

On the one hand we have new untrusted users like NEW PHARMA. For which
we wanted to set to 0 the quantities that shouldn't be ordered.

On the other hand we have orders from ESB. ESB is considered trusted
as it has the duty to do correct checks. And we prefer to have the
errors to be in sync.