Disable sending New Pharma sale order to the ESB
ALCYN-1744

NewPharma creates sale order with the web service directly
sending an increment_id in their call (like the ESB).
So for Odoo it should be a PUT method but the sale order does not
yet exists on the ESB and it fails.
They need to change their calls to resolve this.
They do not connect to Magento.