Add a view for a fast entry of sale lines
When we have more than ~15 lines, the onchanges start to takes a very
long time for each new line (because the onchanges for the other lines
are played again). A view dedicated for the entry of the lines
solves this problem.

Downside is that we need to duplicate the changes in the usual view and
this one.

ALCYN-1708