Add 'base' field on taxes on supplier invoices form
The field was already added manually on the view on production.
The migration script starts by reloading the account invoice views,
in order to remove the manual changes. Otherwise the field would be
duplicated on the view. The script uses convert_file rather than
upgrading the account addon to avoid an upgrade of all the depending
addons.

ALCYN-1582