Remove side-effect in constrains
@api.constrains are for constraints.
If you have any type of error in the constrains, odoo will catch it (it
catches Exception) and re-raise it as a ValidationError.
The nasty effect here is that jobs catch the OperationalError and
automatically retry the job. As they are re-raised as ValidationError
by @api.constrains, the job cannot be retried.