Remove propagation of printed on same pickings of a round
This code prevent to confirm the shippings of a round concurrently,
as everytime it will write on the other shippings.

JEB wrote:

I don't remember why I wrote this. But this code has to be removed
because in reality, it's causing lot's of trouble. They sometimes don't
have enough resources to process all the pickings for a customer and
delivery gets stuck due to this because all picking are suddenly marked
as printed.