Catch all kind of exception on shipping transfer
In delivery rounds, when we do the transfer of a shipping, catch all
kind of exception even unhandled ones, but log them.
This will avoid to have users panicking, but still we need to find and
trace the error, so we'll find the details of the stack trace in the
logs.