Fix wizard for mass sending invoices
ALCYN-1774

When choosing the sending method for an invoice use the sending method
set on the invoice itself which is related to the customer set on the
invoice.
For sending by email the wizard filters out invoices whose commercial
partner does not have an email, but the mail_template uses the email of
the partner_id of the invoice. So fix this filtering.