Fix stock_tests
test_stock_inventory.test_get_products_daily_inventory was failing.
It comes from the fact that it was dependent on the actual year. As
One of the two dates was on a weekend, it could not create an inventory.
I'm using freezegun now in order to set and switch dates.