Fix customer reference on delivery note filename
ALCYN-1802

Using the reference of the partner linked to the stock.picking caused
lots of delivery note not being linked to the proper customer on
Magento.
This is made worse by a specific change that makes a
res_partner ref distinct between a contact and it's addresses.
This change will fix the customer reference for most of the file name
but for some edge cases with the 'client chapeau'.