Fix customer reference on delivery note filename
ALCYN-1802

This improve the previous fix tagged on the same Jira card.
This is done by searching for the customer on a sale order related to the stock picking.
There could be many sale orders and so more than one customer, because
of the way the delivery are grouped.
But this has been discussed as the best solution for the edge case
raised by the 'client chapeau'.