Assign moves with different carriers in different pickings
This is related to "special" delivery methods having dedicated delivery rounds,
such as a Counter or a long-term delivery round.

The previous behavior is:

* When a sale is confirmed with a special delivery, its move are never
grouped with a prior picking, they have the default behavior of creating
a picking per sale.
* The moves of other sales where created in an existing picking with the
same characteristics (location, partner, ...) not yet shipped (if any
otherwise in a new fresh picking).

This is working up to some limits:

* If a picking is created for a special delivery and is still open, then
new "regular" moves will be created in the special picking (and special
delivery round) instead of in a picking assigned to the usual delivery
round of the customer
* If a customer makes a dozen of Counter orders, they will have as much
pickings when they could have only one

The change is to add the delivery method in the grouping of moves per
pickings. As the carrier is not known when "StockMove.assign_picking"
is called, we store in the procurement.group, a group for a sale will
always have its moves grouped with the same delivery method.

ALCYN-1675