Improve handling of groupbypartner with delivery rounds
This is the second part of 4d0e592.

When the delivery round is changed directly on a picking using the
wizard, new moves would still be added to this picking. Now,
the carrier on the procurement group is set on a special
delivery.carrier when the wizard is used. As this delivery method
cannot be selected on sale orders (inactive), moves will never
be added to pickings which have a procurement group using it.

This delivery method can still be selected manually on the procurement
group if needed, if selected, it makes all the transfers of the group
standalone, meaning no new move will be added to them.

This commit also add a post-migration to set the manual delivery method
on every procurement group for "comptoir" and "long-terme" delivery
rounds.

ALCYN-1850