Remove useless new cursor, simplify code
As now the job _job_invoices_by_partners runs only on one partner at
a time, opening a cursor to commit after each partner is useless,
it will be committed at the end of the job.

Take the opportunity to rename the job and adapt the argument to make it
clear it works on one partner. It is safe to forward only the first
partner of the argument to the new job because in the last release, the
code was modified to always pass a list with one partner, and all jobs
of this kind are done on production.