Fix unavailable qty for additional product
Unavailable quantity was copied from line with free products
Recompute it taking the sum of products plus additional products in
mind.

We consider that quantity is reserved first for the ordered product
then for the additional products.

For instance for a remaining quantity of 42, on 40 ordered products
with 1 additional for each 5 ordered products:
 - ordered: 40 -> BO: 0
 - additional: 10 -> BO: 8

ALCYN-1861