Add module web_cache_name_get
We spotted a lot of requests to name_get in the logs.
This module is not likely to fix the problem that we still have to
locate, but it will improve user experience and load in any case.
A use case where it is very noticeable is on a PO or SO with dozens of
lines and a similar tax: the original code will issue a request for each
line to get the name of the same tax, where 1 request was enough.

On my local db, to open a PO with 60 lines with the same tax, the load
time was reduced by about 3 seconds (~6 before 3 after). And 60 name_get
on account.tax were reduced to 1.

BIZ-2823