Optimize number of HTTP requests sent by return wizard
The Return wizard on stock pickings populates a m2m widget with each
line having a many2one to a product of the picking.

These lines are set with the default_get method, at this point nothing
is in database yet, and so the web client will issue one HTTP request
on /web/dataset/call_kw/product.product/name_get per line. Then when
a user edits the quantities, 2 name_get will be send per edit.

So with a picking of 300 lines, the server is spammed with a burst of
300 requests, then couples of requests every edit.

This commit replaces the many2one by a Char field so we set the product
name server-side once.

BIZ-2823