Fix supplier promotion created twice on Magento
ALCYN-1878

Change on supplier promotion are send to the ESB by XML node containing
the action (create, delete) and a checksum that identify the promotion.

Fot this to work the checksum needs to be computed (among others fields)
with the id of the product_supplierinfo record so the checksum is the
same for action on the same promotion. And the ESB will not duplicate
those promotions.