Add fast view for supplier invoice lines
Same as fast view for sale order lines (ALCYN-1708)

When using a one2many widget with >15 lines the view very slow.
This is due to onchanges that are computed for all the lines each
one line is modified. Which makes the view exponentially slower
depending on the number of line.

Using a dedicated list view to create those lines allow to bypass this issue.

ALCYN-1877