Fix order_date on sale order created by the ESB
ALCYN-1930

For sale order created through the web service the date of the order was
set to the moment the order was created in Odoo.
But there is actually a field date that is required in the data for a
call to the web service to be a valid.
So now it is used to set the order_date on the newly created sale order.
This would prevent possible error on price calculation if the
syncrhonisation of sale order is disrupted for days.