ALCYN-1933: fix bug in delivery slip report
and in CSV delivery note.

When a packop serves > 1 stock.move in a picking, the quants are now correctly
distributed among the stock moves. However, when checking for the quants using
move.linked_operation_ids.reserved_quant_ids, the quants mentionned can be on
the operation for another move.

To work around this, we use the straightforward move.quant_ids to get the moved
quants.