Change location filter on stock history by VLB locations
The stock_history view filters by default on internal and transit
usages. This is inaccurate for Alcyon, where it might 'view' as well.
So the report now filter all children of the VLB location, which
really represents what is "internal".

We include the 'Output' location too, because it is not a child of VLB
but should be part of the valorization.

The product_supplier_id was added in a previous version of the view.

ALCYN-1903