ALCYN-1957: Reduce the time to open the list of purchase orders
The field 'nbr_lines_bo' on PO was quite big to compute (itself based on the
'immediately_usable_qty' computed field of product).
As the this field is displayed on the PO form, it has been removed from the
tree view to reduce drastically the time to open the view.