[FIX] wrong computation of average price
The average price was wrongly computed in some cases because the
price on the purchase order was changed *after* the confirmation,
which is the moment when the price used for the computation of the
average price is recorded on the stock move.

We fix this by forcing the recomputation of the price when the move
is received, unconditionnally (normally this is done only when there
is a currency rate conversion).

fix ALCYN-2002