Optimize "Remains to deliver" filter on sale.order.line
The domain is using search filter with dotted path, which is translated
in 3 queries, one getting the ids of orders "is_consignment is false", one
getting the product of type [consu or product] and one searching the
sale order lines, with "product_id in (ids of 1st query) AND order_id in
(ids of 2nd query)". The list of ids of products and orders are almost
the entire tables.

Note: the sale.order.line view is still slow because of the computed
field current_product_qty_unavailable, which will be handled separately.

Note for the release: the migration takes ~8 minutes on my computer.

ALCYN-2008