Fix migration error on res.partner
When updating only account_mass_reconcile, an error is raised with a SQL request
on res_partner having field payment_next_action_date in the select while it
does not exist in database.
As both account_reports and account_credit_control define this field, and as
both modules are installed, we update it to fix the issue.