Add stock_location_act_as_view to flag 'view' locations
Context

The stock locations with a usage == 'view' are not well supported in
Odoo. Pretty often, the code of the stock addon compares the usage with
'internal' and there is no defined behavior when the usage is view,
leading to unintended behavior such as unbalanced quants.

The local code has some logic depending on the locations being views,
but we need the usage to be 'internal'... So, add a new boolean flag
'act_as_view', which is now used for the local logic.

Changes

* Replace location usage by 'internal' in sample/install
* Adapt code using 'usage=view' to work with 'act_as_view' instead
* Migrate current 'view' locations to 'internal' + 'act_as_view'

ALCYN-2043