Refactor stock_inventory_controller
Move the error handling in 'StockInventory.action_done()', because this
is really the whole part we want to rollback on error during
'StockInventory.post_inventory()'. The intention seems clearer.

The previous implementation had a problem when the inventory lines were
created in the same transaction than the one where they are confirmed,
as the wizard to change the qty on hand of a product is doing for
instance. In case of failure, it would crash with a "record not found
error". Here, we ignore lines that don't exist in the new transaction as
anyway they will be rollbacked as well.

The previous implementation was buggy when the same product was used in
several inventory lines: as the line was found with a search on the
same product than the move line, if we had one valid line and an invalid
one, both would be invalid. This is fixed by adding a new Many2one
referencing the inventory line that created the move line. Note: if an
inventory is started before we add the field, we will not be able to
find the line so in this case we raise an error. We could write a
migration but it would as imperfect as the previous solution as we could
only rely on the identical product.

The try-except on MissingError seems useless (no reason to happen) and
indeed, I could find zero occurrences of this error in the inventory
lines in production: removed it.