ALCYN-2070 fix last historic average price of products
some products have historic prices which are wrong because the PO
was modified after confirmation. We do not fix the history, rather
we change the last historic price to the price of the lastest purchase
for the product.

The intention is that this price is likely closer to the real average
price of the product than the computed one, and this will help
the average price to converge back to a better value in the future.