Fix stock history MV view category and suppliers mixup
This view uses an UNION ALL. In the 2 SELECT combined together, the
fields product_categ_id and product_supplier_id were not in the same
order, leading to inversion of the ids (id of the supplier in the
category and conversely).

For users, it leds to incoherent results in the Stock History view when
using the groups by category.

ALCYN-2124