Fix product used to add shipping cost on order line
Up to now the product used to create a sale order line  for the specific
Alcyon shipping costs was hardcoded and the xmlid used was wrong.
So by using the product declared on the delivery carrier the problem is
fixed and also allows for configuration of it in the future.

ALCYN-2094