Fix MemoryError on WS sale order creation
base_exception module had a default behavior of running the exception
check on all records of a model if the check is called with
an empty `self`.

Seems we reached the memory limit due to the 3M of sale order lines.

When creating sale order throught WS we prepare the data before creation
using onchanges, I suspect that at some point we trigger the exception
checks in a api.constraint due to the onchanges while self is still
unset.

ALCYN-2195