Remove side-effect in constrains
@api.constrains are for constraints solely.

If you have any type of error in the constrains, odoo will catch it (it
catches Exception) and re-raise it as a ValidationError.
The nasty effect here is that jobs catch the OperationalError and
automatically retry the job. As they are re-raised as ValidationError
by @api.constrains, the job cannot be retried. HTTP dispatch suffers
from the same issue.

A last occurrence of this issue remains, which will be fixed in
https://github.com/camptocamp/alcyon_odoo/pull/1541 as the task
ALCYN-2091 needed rework on this part.

ALCYN-2201