Fix concurrent transactions issues on sale_exception
* Description and discussion of the issue https://github.com/OCA/server-tools/issues/1642
* Include PR https://github.com/OCA/server-tools/pull/1644
* Include PR https://github.com/OCA/sale-workflow/pull/916
* Include PR https://github.com/OCA/server-tools/pull/1647
* Remove https://github.com/OCA/server-tools/pull/1620 which has been
merged
* Remove https://github.com/OCA/server-tools/pull/1638, replaced by #1647

Fix concurrent transactions issues on sale_exception
----------------------------------------------------

(PR server-tools#1642 + sale-workflow#916)

The method called by 'sale_check_exception' has a side effect, it writes
on 'exception.rule' + on the Many2many relation between it and
sale.order(.line). When decorated by @api.constrains, any error during
the method will be caught and re-raised as "ValidationError".
This part of code is very prone to concurrent updates as 2 sales having
the same exception will both write on the same 'exception.rule'.
A concurrent update (OperationalError) is re-raised as ValidationError,
and then is not retried properly.

Calling the same method in create/write has the same effect than
@api.constrains without shadowing the exception type.

OperationalError shadowed by ValidationError in api.constrains
--------------------------------------------------------------

(PR server-tools#1647)

The goal of the modified method is to create or remove the relationship
(in the M2m relation tabel) between the tested model (such as
sale_order) and the exception rules. When the ORM writes on
ExceptionRule.sale_ids (using the example of sale_exception), it will
first proceeds with these updates:

* An UPDATE on exception_rule to set the write_date
* INSERT or DELETE on the relation table
* But then, as "write" is called on the exception rule, the ORM will
  trigger the api.depends to recompute all the "main_exception_ids"
  of the records (sales, ...) related to it, leading to an UPDATE
  for each sale order

We end up with RowExclusiveLock on such records:

* All the records of the relation table added / deleted for the current
  sale order
* All the records of exception_rule matching the current sale order
* All the records of sale_order related to the exception rules matching
  the current sale order

The first one is expected, the next 2 are not. We can remove the lock on
the exception_rule table by removing _log_access, however in any case,
the main_exception_ids computed field will continue to lock many sale
orders, effectively preventing 2 sales orders with the same exception
to be confirmed at the same time.

Reversing the write by writing on SaleOrder instead of ExceptionRule
fixes the 2 unexpected locks. It should not result in more queries: the
"to remove" part generates a DELETE on the relation table for the rule
to remove and the "to add" part generates an INSERT for the rule to add,
both will be exactly the same in both cases.