Fix total amount of discounts in invoice reports
The value was rounded before being multiplied by the quantity, leading
to differences between the computed discount amount and the expected
one. The rounding is not even necessary as assigning the value to
the Monetary fields will automatically round it with the field's
precision.

I took the opportunity to remove the (1 - (1 - x) as it's equal to x.

ALCYN-2214