Add product_last_transaction
When computing an inventory to date (stock.history model)
there is two fields available when exporting the data
(last purchasing date, last selling date).
Those fields are also needed when exporting an inventory valuation
(stock.quant model)
For this a new module is added that move the computation on the
product.product model.

ALCYN-2129