Fix differences with inventory to date and valuation
The computation of the stock at a specific date based on the
stock.history table gives some eroneuous information compare to the valuation
one done on the stock.quant table.
One of the reason is that the computation at date is done from the past
to now.
Here the computation at a past date is done starting from the now time.
To do so for each product a record is added in the hsitory table with
the current stock state.

ALCYN-2236

[IMP] simplify location selection

use suggestion from jbaudoux