[FIX] failing tests in specific_report
the module account_invoice_triple_discount is not playing nice with us, and will sometimes
not work at all. We fix this in https://github.com/acsone/account-invoicing/pull/2 by detecting
the problem and recomputing.

I tested this by running the unit tests of specific_report 100x: without the fix they crash
abount 1/30 times, with the fix everything is ok.