Fix cancellation of sale orders after delay
The job takes an argument 'creation_date'. We expect the value to be
the datetime when the controller calls Odoo. But the variable was
initialized with 'datetime.now()' at the execution of the job.
Pass the parameter creation_date at scheduling of the job to have
the expected datetime.

Note: for backward compatibility with jobs scheduled before the
deployment, we have to keep the creation_date parameter optional.
But it should be made mandatory with a positional argument later.

ALCYN-2309