[FIX] ALCYN-2317
don't force the update of price_unit_base when price_unit is changed, as
this defeats the whole purpose of having 2 fields, one containing the
discounted price (price_unit) and one for the displayed not discounted
price (price_unit_base).

Fix the test by updating the correct field (price_unit_base) and
checking that the onchange and call to write work properly.