ALCYN-2150 implement a false MTO behavior
When a product with the MTO route is sold, a resupply can be triggered.

Differences from the standard MTO route:

* the resupply moves are not chained to the delivery moves

Differences from the Alcyon MTS mode:

* the supplier management days are not taken into account and a RFQ can
be generated immediately

This is achieved by overriding sale_order.action_confirm:
* we check for products with route MTO
* we make sure these products have an stock orderpoint (if required, we
create one, with min/max = 0/0)
* for these products, we force an immediate orderpoint recomputation

We changed the way the orderpoint recomputation is performed so that in
that case, we don't take the suppliers management days.