specific_sale: Improve _compute_production_lot_ids perf on SO
Refactor the computation of the production lots on the sale order lines to be batch computed.
In the same time, extract and rewrite the computation of the available qty on the lot to also be batch computed into a new addon acl_stock_lot_available (+ add tests).

Number of SQL queries required to compute the lots on a so:
 * Before: nbr lines * nbr lots
 * After: 2 queries.

On a user experience point of view, before this change, the deletion of a line into a sale order could take more than 30 secs on a so with 25 lines. Now it takes less than 3 secs.