ALCYN-2266 / ALCYN-2401 restore locking
I spent a lot of time investigating this, with no clear answer. I
finally came up with this commit at a workaround. It fixes the issue of
deadlocking when a round is delivered. There may be other cases of
deadlocking we have not found out yet.

The deadlocking comes from a transaction lock on the round being
delivered (related to the database insulation level) and the manual
locks we are taking, and possibly the new cursor which is created for
the error processing during delivery. I could find no easy way to
guarantee that the locks would be taken in the same order always without
locking the round during the delivery of a customer instance. I think
the issue comes from the ALCYN-2130 but I'm not 100% sure.

* to be cleaner, don't pass a recordset in the context
(backorder_assign): this can be in a different environment than self
at runtime because the delivery job will create a new environment for
error processing.
(this is not the source of the deadlocking, though :-( )

* add a lock on the round instance being delivered in the customer
instance delivery jobs. This has the net effect of serializing these
jobs, which means the whole delivery process takes longer overall, but
at least we don't get the deadlock.