ALCYN-2419: fix case 17C (edge case of ALCYN-2130)
Added 2 automated tests for cases 17B and 17C.

Test case 17B passed without change, this was confirmed by JBX.

For case 17C, added a context key to be able to tell apart the creation
of a backorder by a partial picking from the creation of a backorder
because we are in the process of delivering a round instance
(`detaching_from_round`). In the latter case, when a new picking is
created, added some code to try and find a round instance to which the
picking can be added.

To make it work, fix some issues:

* change the state of the round instance to 'delivering' before running
the delivery of the customer rounds (only a problem in the tests where
we deliver synchronously)

* unlink the linked_move_operation_ids of the move which is assigned to
a new picking before the assignment, otherwise they are still there and
the move stays linked to an operation which is in the wrong stock
picking, and this prevents assigning the new picking to a new round